Add random background functionality
It uses the Pexels API to retreive images using the search term 'Japan'.
It will pick a random page number from the results, then it will get a random index
from the photos array. From there, it will set the background of the body element.
At the moment, the images load slowly.