Add model for the stack and a stack handler
There's a bit more added than that. I modified how the
program is going to work. Basically all of the information
is going to be saved in the stack model, and most of
the manipulation will be done on the stack. The stack
handler will hold the functions that will manipulate
the stack. Along with that, I started working on the logic
for showing answers and answering. At the moment, the logic
for missing an answer works fine. I have to test it all
out once I enable the different stack options.