Rework the move character function
I wanted this function to be used for both the hiragana and katakana button,
so I reworked it to be based on the option and character set passed to it. Also,
to help with this I removed the shownCharacters array. That way, it only ever
has to move characters between 2 arrays. Plus, I realized that the shownCharacters
array wasn't needed. I could just move the characters to the end of the array
instead of to an entirely different array. This also gives me the option to change
how characters are moved within the characters array after the user gives selects
an answer.