Add functionality to the random button in settings
This was easy, but I was modified how the background was loaded. I randomly
tested to see if it would load a portrait image, and it didn't. I realized
one of the URLs had the orientation set to landscape. I changed that to use
the getOrientation function and it works fine now.